Move buildLinks and buildNodes to graph.builder.js
