+++ src/components/graph/graph.renderer.jsx
@@ -11,7 +11,7 @@ import { MARKERS, MARKER_SMALL_SIZE, MARKER_MEDIUM_OFFSET, MARKER_LARGE_OFFSET }
 import Link from "../link/Link";
 import Node from "../node/Node";
 import Marker from "../marker/Marker";
-import { buildLinkProps, buildNodeProps } from "./graph.helper";
+import { buildLinkProps, buildNodeProps } from "./graph.builder";
 import { isNodeVisible } from "./collapse.helper";
 
 /**
